requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.1.0
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorlog==6.7.0
constantly==15.1.0
cron-descriptor==1.4.0
cryptography==41.0.1
Cython==0.29.35
daphne==4.0.0
decorator==4.4.2
deffcode==0.2.5
Deprecated==1.2.14
Django==4.2.2
django-appconf==1.0.5
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.1.0
django-debug-toolbar==4.1.0
django-elasticsearch==7.0.2
django-elasticsearch-dsl==7.3
django-elasticsearch-dsl-drf==0.22.5
django-environ==0.10.0
django-hashid-field==3.3.7
django-nine==0.2.7
django-otp==1.2.2
django-redis==5.3.0
django-redis-sessions==0.6.2
django-request-token==2.2
django-rest-encrypted-lookup==0.10.1
django-timezone-field==5.1
django-versatileimagefield==3.0
django-video-encoding==1.0.0
django-webpush==0.3.5
djangorestframework==3.14.0
djangorestframework-api-key==2.3.0
djangorestframework-simplejwt==5.2.2
dnspython==2.3.0
elasticsearch==7.17.9
elasticsearch-dsl==7.4.1
eventlet==0.33.3
exceptiongroup==1.1.1
ffprobe-python==1.0.3
flake8==6.0.0
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
h2==4.1.0
hashids==1.3.1
hpack==4.0.0
http-ece==1.1.0
hypercorn==0.14.3
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
incremental==22.10.0
iniconfig==2.0.0
kombu==5.3.1
mccabe==0.7.0
moviepy==1.0.3
numpy==1.25.0
opencv-python==4.7.0.72
packaging==23.1
Pillow==9.5.0
pluggy==1.0.0
priority==2.0.0
proglog==0.1.10
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
py-vapid==1.9.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser==2.21
pycrypto==2.6.1
pyflakes==3.0.1
PyJWT==2.7.0
pyOpenSSL==23.2.0
pyotp==2.8.0
pyparsing==3.1.0
pytest==7.3.2
pytest-django==4.5.2
pytest-sugar==0.9.7
python-crontab==2.7.1
python-dateutil==2.8.2
python-magic==0.4.27
pytz==2023.3
pywebpush==1.9.4
redis==4.5.5
requests==2.31.0
scikit-video==1.1.11
scipy==1.10.1
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
termcolor==2.3.0
thumbnail==1.5
tinytag==1.9.0
toml==0.10.2
tomli==2.0.1
tqdm==4.65.0
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
uvicorn==0.22.0
uvloop==0.17.0
vidgear==0.3.0
vine==5.0.0
Wand==0.6.11
wcwidth==0.2.6
wrapt==1.15.0
wsproto==1.2.0
zope.interface==6.0
